+++ web/webkit/src/main/scala/net/liftweb/http/LiftRules.scala
@@ -897,26 +897,6 @@ class LiftRules() extends Factory with FormVendor with LazyLoggable {
    */
   val dataAttributeProcessor: RulesSeq[DataAttributeProcessor] = new RulesSeq()
 
-  private def makeMetaData(key: String, value: String, rest: MetaData): MetaData = key.indexOf(":") match {
-    case x if x > 0 => new PrefixedAttribute(key.substring(0, x),
-      key.substring(x + 1),
-      value, rest)
-
-    case _ => new UnprefixedAttribute(key, value, rest)
-  }
-
-  private def pairsToMetaData(in: List[String]): MetaData = in match {
-    case Nil => Null
-    case x :: xs => {
-      val rest = pairsToMetaData(xs)
-      x.charSplit('=').map(Helpers.urlDecode) match {
-        case Nil => rest
-        case x :: Nil => makeMetaData(x, "", rest)
-        case x :: y :: _ => makeMetaData(x, y, rest)
-      }
-    }
-  }
-
   dataAttributeProcessor.append {
     case ("lift", snippetInvocation, element, liftSession) =>
       snippetInvocation.charSplit('?') match {
